requirements.txt
numpy
cython==0.29.33
scikit-image
scipy
scikit-learn
smplx==0.1.28
tqdm
yacs
numba
opencv-python
tensorboardx
filterpy
cython
chumpy
Pillow
trimesh
pyrender
matplotlib
json_tricks
torchgeometry
einops
joblib
boto3
requests
easydict
pycocotools
plyfile
timm
pyglet
mmdet==2.26.0